cambios bodega , producto , relleno , compra , tienda ,funciones , services

+++ Ventas/backend/src/entity/tienda.entity.js
@@ -16,19 +16,19 @@ const TiendaSchema = new EntitySchema({
       nullable: false,
     },
     medida_ancho: {
-      type: "int",
+      type: "numeric",
       precision: 10,
       scale: 2,
       nullable: true,
     },
     medida_espesor: {
-      type: "int",
+      type: "numeric",
       precision: 10,
       scale: 2,
       nullable: true,
     },
     medida_alto: {
-      type: "int",
+      type: "numeric",
       precision: 10,
       scale: 2,
       nullable: true,
@@ -62,5 +62,6 @@ const TiendaSchema = new EntitySchema({
     },
   ],
 });
+
 
 export default TiendaSchema;
